+++ packages/components/slider/src/Slider.styles.ts
@@ -1,71 +1,55 @@
-import styled from 'styled-components';
+import styled, { css, FlattenSimpleInterpolation } from 'styled-components';
 import Slider, { SliderProps } from 'antd/lib/slider';
 import { ThemeProps } from '@synerise/ds-core/dist/js/DSProvider/ThemeProvider/theme';
 import { ComponentType } from 'react';
 
-const multipleTracksStyles = (props: ThemeProps): string => `
-  .ant-slider-track {
-    &-1,
-    &-11 {
-      background-color: ${props.theme.palette['green-600']};
-    }
+export declare type colorMapProps = {
+  [key: string]: string;
+};
 
-    &-2,
-    &-12 {
-      background-color: ${props.theme.palette['cyan-600']};
-    }
+const indexMap = {
+  '0': 'green-600',
+  '1': 'cyan-600',
+  '2': 'fern-600',
+  '3': 'mars-600',
+  '4': 'orange-600',
+  '5': 'pink-600',
+  '6': 'purple-600',
+  '7': 'red-600',
+  '8': 'violet-600',
+  '9': 'yellow-600',
+};
 
-    &-3,
-    &-13 {
-      background-color: ${props.theme.palette['fern-600']};
-    }
-
-    &-4,
-    &-14 {
-      background-color: ${props.theme.palette['mars-600']};
-    }
-
-    &-5,
-    &-15 {
-      background-color: ${props.theme.palette['orange-600']};
-    }
-
-    &-6,
-    &-16 {
-      background-color: ${props.theme.palette['pink-600']};
-    }
-
-    &-7,
-    &-17 {
-      background-color: ${props.theme.palette['purple-600']};
-    }
-
-    &-8,
-    &-18 {
-      background-color: ${props.theme.palette['red-600']};
-    }
-
-    &-9,
-    &-19 {
-      background-color: ${props.theme.palette['violet-600']};
-    }
-
-    &-10,
-    &-20 {
-      background-color: ${props.theme.palette['yellow-600']};
-    }
-  }
-`;
+const createTracksStyles = (props: ThemeProps, colorsMap: colorMapProps): FlattenSimpleInterpolation => {
+  const styles = Object.values(colorsMap).map(
+    (color: string, index: number) => `
+      &-${index + 1},
+      &-${index + 11} {
+        background-color: ${props.theme.palette[color]};
+      }
+  `
+  );
+  const style = styles.join('');
+  return css`
+    ${style}
+  `;
+};
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const AntdSlider = styled((Slider as any) as ComponentType<Omit<SliderProps, 'value'>>)<{
   useColorPalette?: boolean;
+  tracksColorMap?: colorMapProps;
 }>`
-  ${(props): string => (props.useColorPalette ? multipleTracksStyles(props) : '')}
-  &.ant-slider:hover {
-    ${(props): string => (props.useColorPalette ? multipleTracksStyles(props) : '')};
+  .ant-slider-track {
+    ${(props): FlattenSimpleInterpolation =>
+      props.useColorPalette ? createTracksStyles(props, props.tracksColorMap ? props.tracksColorMap : indexMap) : css``}
   }
-
+  &.ant-slider:hover {
+    ${(props): FlattenSimpleInterpolation =>
+      props.useColorPalette
+        ? createTracksStyles(props, props.tracksColorMap ? props.tracksColorMap : indexMap)
+        : css``};
+  }
   .ant-tooltip-inner {
     font-size: 13px;
     padding: 3px 7px;