+++ packages/components/navbar/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.2.6...@synerise/ds-navbar@0.2.7) (2020-01-14)
+
+### Bug Fixes
+
+- **navbar:** removed duplicated navbar ([2a7c038](https://github.com/Synerise/synerise-design/commit/2a7c038ea7ebf1460c6f1882cc1992b99f8c168d))
+
 ## [0.2.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.2.4...@synerise/ds-navbar@0.2.5) (2020-01-02)
 
 **Note:** Version bump only for package @synerise/ds-navbar