+++ packages/components/slider/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.4.0...@synerise/ds-slider@0.5.0) (2020-01-13)
+
+### Features
+
+- **slider:** allow add color map for slider ([2b5d6a9](https://github.com/Synerise/synerise-design/commit/2b5d6a940314ac169cd8cca1d97c8ca79f34f76a))
+
+# [0.4.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.3.0...@synerise/ds-slider@0.4.0) (2020-01-13)
+
+### Features
+
+- **slider:** allow add color map for slider ([6d34662](https://github.com/Synerise/synerise-design/commit/6d346624b9c0548add47691a4523998557c39719))
+
+# [0.3.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.6...@synerise/ds-slider@0.3.0) (2020-01-10)
+
+### Features
+
+- **slider:** allow add color map for slider ([ccf6654](https://github.com/Synerise/synerise-design/commit/ccf66542ef732f65fe374ea910de5a882ecb1d16))
+
 ## [0.2.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.4...@synerise/ds-slider@0.2.5) (2020-01-02)
 
 **Note:** Version bump only for package @synerise/ds-slider