+++ packages/portal/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.29.2](https://github.com/synerise/synerise-design/compare/portal@0.29.1...portal@0.29.2) (2020-01-14)
+
+
+### Bug Fixes
+
+* **modal:** fix onClick method on close button ([60d3ce2](https://github.com/synerise/synerise-design/commit/60d3ce2452a3451a69598dc46335e25c134afa80))
+* **navbar:** removed duplicated navbar ([2a7c038](https://github.com/synerise/synerise-design/commit/2a7c038ea7ebf1460c6f1882cc1992b99f8c168d))
+
+
+
+
+
+## [0.29.1](https://github.com/synerise/synerise-design/compare/portal@0.29.0...portal@0.29.1) (2020-01-13)
+
+
+### Bug Fixes
+
+* **list:** add divider list type solid ([83c89bd](https://github.com/synerise/synerise-design/commit/83c89bd8556c9d82602d5f713d42790979a96a0a))
+* **list:** delete wraper on action element ([783624d](https://github.com/synerise/synerise-design/commit/783624dfbfb938eaaa013c396c4959beeea0fee2))
+
+
+
+
+
+# [0.29.0](https://github.com/synerise/synerise-design/compare/portal@0.28.0...portal@0.29.0) (2020-01-13)
+
+
+### Features
+
+* **slider:** allow add color map for slider ([6d34662](https://github.com/synerise/synerise-design/commit/6d346624b9c0548add47691a4523998557c39719))
+
+
+
+
+
+# [0.28.0](https://github.com/synerise/synerise-design/compare/portal@0.27.2...portal@0.28.0) (2020-01-10)
+
+
+### Features
+
+* **slider:** allow add color map for slider ([ccf6654](https://github.com/synerise/synerise-design/commit/ccf66542ef732f65fe374ea910de5a882ecb1d16))
+
+
+
+
+
 ## [0.27.1](https://github.com/synerise/synerise-design/compare/portal@0.27.0...portal@0.27.1) (2020-01-02)